+++ vitrina/datasets/models.py
@@ -167,15 +167,19 @@ class Dataset(Resource):
     PUBLIC = "PUBLIC"
     RESTRICTED = "RESTRICTED"
     NON_PUBLIC = "NON_PUBLIC"
+    CONFIDENTIAL = "CONFIDENTIAL"
+
     ACCESS_RIGHTS = (
-        (PUBLIC, _("Atviri duomenys")),
-        (RESTRICTED, _("Apsaugoti duomenys")),
-        (NON_PUBLIC, _("Uždari duomenys")),
+        (PUBLIC, _("Vieši")),
+        (RESTRICTED, _("Apriboti")),
+        (NON_PUBLIC, _("Nevieši")),
+        (CONFIDENTIAL, _("Konfidencialūs")),
     )
     FILTER_ACCESS_RIGHTS = {
-        PUBLIC: _("Atviri duomenys"),
-        RESTRICTED: _("Apsaugoti duomenys"),
-        NON_PUBLIC: _("Uždari duomenys"),
+        PUBLIC: _("Vieši"),
+        RESTRICTED: _("Apriboti"),
+        NON_PUBLIC: _("Nevieši"),
+        CONFIDENTIAL: _("Konfidencialūs"),
     }
 
     API_ORIGIN = "api"
@@ -320,7 +324,7 @@ class Dataset(Resource):
         choices=ACCESS_RIGHTS,
         max_length=255,
         help_text=_(
-            "Informacija ar duomenų rinkinys yra atviri duomenys, ar jam taikomi prieigos apribojimai, ar jis nėra viešas. Atitinka dct:accessRights."
+            "Informacija ar duomenų rinkinys yra atviri duomenys, ar jam taikomi prieigos apribojimai, ar jis nėra viešas, ar jis konfidencialus. Atitinka dct:accessRights."
         ),
     )